requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
faster-whisper
pydub
pyannote.audio
torch
torchaudio
slowapi>=0.1.9
google-genai>=0.2.0
boto3>=1.34.0
requests>=2.31.0
